command-treatment-factory renamed to delegator (#12)
